Update arr_content.py
Changed the fractional scores for meta-review form options under meta-review button from integer to float.

+++ openreview/stages/arr_content.py
@@ -1637,7 +1637,7 @@ arr_metareview_content = {
                     }
                 ],
                 "optional": False,
-                "type": "integer"
+                "type": "float"
             }
         },
         "order": 4
